introducing apimonitoring reconciler builder, moving integration tests to separate file
